finished token authentication

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -17,8 +18,11 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const users = await User.find({})
-  const user = users[0]
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({ ...body, user: user.id })
   const savedBlog = await blog.save()
@@ -36,6 +40,15 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log(decodedToken)
+  if (!request.token || !decodedToken.id)
+    return response.status(401).json({ error: 'Must be logged in to delete posts' })
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id.toString() !== blog.user.toString())
+    return response.status(401).json({ error: 'Cannot delete blogs that weren\'t posted by you' })
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
